[api] split api and http clients
Redesign the `api` module to dissociate the API from the HTTP client.
Make HTTP clients pluggable.

**Changes**
Two new modules:
* `api_client`.
  * `APIClient`. Defines
* `http_client`.
  * `HTTPClient`. Defines a pluggable HTTP client skeleton.
    * `RequestClient`. HTTP client based on `requests` 3p module.

One module renamed:
* `api.base` module is now `api.resources`
